+++ source/CHANGELOG.md
@@ -5,6 +5,7 @@ Andreas Abel <andreas.abel@gu.se>  (unreleased)
 * Haskell: fixes in layout preprocessor [#343,#344,#345]
 * C: fixed a space leak when parsing from a string in memory [#347]
 * C: new methods `free*` to deallocate syntax trees [#348]
+* C++: `define`d constructors now reside in `Absyn` [#287]
 * Ocaml: fixed translation of nested `define`d constructors
 
 # 2.9.1